requirements.txt
pybit==5.10.1
python-dotenv==1.0.0
requests==2.32.3
websocket-client==1.8.0
plotly==5.24.0
dash==2.17.1
dash-bootstrap-components==1.6.0
psutil==6.1.0